scheduler/share/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
coloredlogs==15.0.1
comm==0.1.3
DateTime==4.4
debugpy==1.6.7
decorator==5.1.1
Django==4.2.3
djangorestframework==3.14.0
dnspython==2.2.1
et-xmlfile==1.1.0
executing==1.2.0
humanfriendly==10.0
idna==3.3
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.12.2
jedi==0.18.2
jupyter-client==8.2.0
jupyter-core==5.3.0
matplotlib-inline==0.1.6
mysql-connector-python==8.0.28
nest-asyncio==1.5.6
numpy==1.23.2
openpyxl==3.0.10
packaging==23.1
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.21.5
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.1
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing-extensions==4.5.0
urllib3==1.26.9
wcwidth==0.2.6
zipp==3.15.0
zope.interface==5.4.0
